+++ lib/msf/base/sessions/command_shell.rb
@@ -759,25 +759,5 @@ protected
   end
 end
 
-class CommandShellWindows < CommandShell
-  def initialize(*args)
-    self.platform = "windows"
-    super
-  end
-  def shell_command_token(cmd,timeout = 10)
-    shell_command_token_win32(cmd,timeout)
-  end
-end
-
-class CommandShellUnix < CommandShell
-  def initialize(*args)
-    self.platform = "unix"
-    super
-  end
-  def shell_command_token(cmd,timeout = 10)
-    shell_command_token_unix(cmd,timeout)
-  end
-end
-
 end
 end
