+++ lib/msf/ui/console/command_dispatcher/payload.rb
@@ -1,6 +1,7 @@
 # -*- coding: binary -*-
 
 require 'rex/parser/arguments'
+require 'msf/util/exe'
 
 module Msf
   module Ui
