+++ lib/msf/ui/console/module_command_dispatcher.rb
@@ -1,5 +1,6 @@
 # -*- coding: binary -*-
 require 'msf/ui/console/command_dispatcher'
+require 'rex/parser/arguments'
 
 module Msf
 module Ui
