+++ tools/exploit/egghunter.rb
@@ -13,7 +13,6 @@ $:.unshift(File.expand_path(File.join(File.dirname(msfbase), '..', '..', 'lib'))
 require 'msfenv'
 require 'rex'
 require 'optparse'
-require 'rex/exploitation/egghunter'
 
 module Egghunter
   class OptsConsole
@@ -128,7 +127,7 @@ module Egghunter
         list_formats
         return
       end
-
+      require 'rex/exploitation/egghunter'
       egghunter = Rex::Exploitation::Egghunter.new(@opts[:platform], @opts[:arch])
       raw_code = egghunter.hunter_stub('', @opts[:badchars], @opts)
       output_stream = $stdout
